safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/deflate-frame-invalid-parameter-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/deflate-frame-invalid-parameter?x-foo' failed: Received unexpected deflate-frame parameter
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/deflate-frame-invalid-parameter?max_window_bits=7' failed: Received invalid max_window_bits parameter
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/deflate-frame-invalid-parameter?max_window_bits=16' failed: Received invalid max_window_bits parameter
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/deflate-frame-invalid-parameter?no_context_takeover=foo' failed: Received invalid no_context_takeover parameter
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/deflate-frame-invalid-parameter?max_window_bits=8;%20no_context_takeover;%20x-foo' failed: Received unexpected deflate-frame parameter
Test whether WebSocket rejects invalid deflate-frame parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing parameter: "x-foo"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=7"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=16"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "no_context_takeover=foo"
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "max_window_bits=8; no_context_takeover; x-foo"
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
